allow s_client(1) to fail
When using s_client(1), it is possible for the
connection to fail to be established, but the list of
ciphers still to be determined, for example if a
server requires client authentication.

Hence, parse output slightly differently to ensure
that established connections are caught even if the
connection is then terminated.